requirements.txt
dash==2.17.1
dash_bootstrap_components==1.6.0
dash_daq==0.5.0
grpcio==1.67.0
grpcio-tools==1.67.0
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
plotly==5.22.0
protobuf==5.28.3
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
tensorboard
